moved the sparkle into a new component:Sparkle
