+++ src/components/Homepage/HomepageStepper/Step/Sparkle/Sparkle.css
@@ -1,0 +1,20 @@
+.Sparkle {
+  position: absolute;
+  height: 6px !important;
+  width: 6px;
+  background-color: rgba(255, 255, 255, 0.658);
+  border-radius: 50%;
+  animation: sparkle 3s ease-in-out infinite;
+}
+
+@keyframes sparkle {
+  0%,
+  100% {
+    opacity: 0;
+    transform: scale(0.5);
+  }
+  50% {
+    opacity: 1;
+    transform: scale(1);
+  }
+}
