+++ src/components/Homepage/HomepageStepper/Step/Step.css
@@ -1,15 +1,3 @@
-@keyframes sparkle {
-  0%,
-  100% {
-    opacity: 0;
-    transform: scale(0.5);
-  }
-  50% {
-    opacity: 1;
-    transform: scale(1);
-  }
-}
-
 .Step {
   position: absolute !important;
   display: flex;
@@ -57,15 +45,6 @@
   border-radius: 50% !important;
 }
 
-.Sparkle {
-  position: absolute;
-  height: 6px !important;
-  width: 6px;
-  background-color: rgba(255, 255, 255, 0.658);
-  border-radius: 50%;
-  animation: sparkle 4s ease-in-out infinite;
-}
-
 .Golden {
   background: linear-gradient(
     130deg,
@@ -92,7 +71,6 @@
   opacity: 0;
 }
 
-/* Step Icon Styles */
 .StepIcon {
   font-size: 25px !important;
   color: var(--text) !important;
@@ -123,8 +101,8 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  width: 125px !important;
-  height: 75px !important;
+  width: 100px !important;
+  height: 50px !important;
   background-color: var(--first-color) !important;
   background: linear-gradient(
     130deg,
