+++ src/components/Homepage/HomepageStepper/Step/Step.tsx
@@ -5,7 +5,8 @@ import { Box, Button } from "@mui/material";
 import FitnessCenterIcon from "@mui/icons-material/FitnessCenter";
 import CheckIcon from "@mui/icons-material/Check";
 import { Session } from "../../../../types/Session";
-import { useEffect, useState } from "react";
+import { CalcStepPosition } from "../../../../helperFuncs/CalcStepPosition";
+import Sparkle from "./Sparkle/Sparkle";
 
 type StepProps = {
   session: Session;
@@ -14,40 +15,28 @@ type StepProps = {
   currentSessionId: number | null;
 };
 
-type Position = "left" | "top";
-
 function Step({
   session,
   sessionNum,
   amountOfSteps,
   currentSessionId,
 }: StepProps) {
-  const [randomPos, setRandomPos] = useState("");
-
-  const CalcPosition = (position: Position) => {
-    if (position === "left")
-      return `${
-        42 + 20 * Math.sin((2 * Math.PI * sessionNum) / amountOfSteps)
-      }%`;
-    else if (position === "top") return `${sessionNum * 14.7}%`;
-  };
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setRandomPos(`${Math.random() * 50}px`);
-    }, 4000);
-
-    return () => clearInterval(interval);
-  }, []);
-
   return (
     <>
       {currentSessionId === session.id && (
         <Box
           className="StartContainer"
           sx={{
-            top: `calc(${CalcPosition("top")} - 115px)`,
-            left: `calc(${CalcPosition("left")} - 65px)`,
+            top: `calc(${CalcStepPosition(
+              "top",
+              sessionNum,
+              amountOfSteps
+            )} - 95px)`,
+            left: `calc(${CalcStepPosition(
+              "left",
+              sessionNum,
+              amountOfSteps
+            )} - 50px)`,
           }}
         >
           <Button className="StartContainerRectangle">Start</Button>
@@ -57,18 +46,13 @@ function Step({
         key={session.id}
         className={!session.isGolden ? "Step" : "Step Golden"}
         sx={{
-          top: CalcPosition("top"),
-          left: CalcPosition("left"),
+          top: CalcStepPosition("top", sessionNum, amountOfSteps),
+          left: CalcStepPosition("left", sessionNum, amountOfSteps),
         }}
       >
         <Box className="StepContainer">
           {session.type}
-          {session.isGolden && (
-            <Box
-              className="Sparkle"
-              sx={{ top: randomPos, left: randomPos }}
-            ></Box>
-          )}
+          {session.isGolden && <Sparkle />}
           {session.isComplete ? (
             <CheckIcon className="StepIcon" fontSize="large" />
           ) : (
